pretty much all: get rid of renderables and use primitives
By doing this, we allow the Vec of primitives to be immediatly usable
as kernel parameters. This has 2 major advantages:
1. the renderable data doesn't need to be gathered for every frame, which
gets rid of a lot of copying data on the cpu side
2. the data is uploaded as 1 struct, which means that we can use a struct with
the same data structure in the kernel and access parameters in that struct instead
of fiddling around with multiple arrays, having to calculate at what index a certain
parameter of an object is in the array. It makes it much more readable and maintainable.